Update Layout Component and Add Menu Component
