+++ src/components/Layout/Layout.component.jsx
@@ -1,42 +1,21 @@
 import React, { useState } from 'react';
-import {
-  IoEnterOutline,
-  IoExitOutline,
-  IoHomeOutline,
-  IoLogoYoutube,
-  IoOptions,
-  IoStar,
-} from 'react-icons/io5';
-import { useHistory } from 'react-router';
-import { useAuth } from '../../providers/Auth';
-import Searchbar from '../Searchbar/Searchbar.component';
+import { IoLogoYoutube } from 'react-icons/io5';
+import Menu from '../Menu';
+import Searchbar from '../Searchbar';
 import UserImage from '../UserImage';
 import {
   Container,
   Header,
-  Menu,
-  MenuList,
-  MenuItem,
-  MenuLink,
   HeaderLeft,
   HeaderRight,
   HeaderCenter,
   MainContent,
-  MenuIcon,
   MenuHamburger,
   Logo,
 } from './Layout.styles';
 
 const Layout = ({ children }) => {
   const [isMenuExpanded, setIsMenuExpanded] = useState(false);
-  const history = useHistory();
-  const { authenticated, logout } = useAuth();
-
-  const deAuthenticate = (event) => {
-    event.preventDefault();
-    logout();
-    history.push('/');
-  };
 
   const toggleMenu = () => {
     setIsMenuExpanded(!isMenuExpanded);
@@ -63,52 +42,7 @@ const Layout = ({ children }) => {
           <UserImage />
         </HeaderRight>
       </Header>
-      <Menu data-testid="menu">
-        <MenuList>
-          <MenuItem>
-            <MenuLink to="/" exact>
-              <MenuIcon isMenuExpanded={isMenuExpanded}>
-                <IoHomeOutline />
-                <span>Home</span>
-              </MenuIcon>
-            </MenuLink>
-          </MenuItem>
-          <MenuItem disabled>
-            <MenuIcon isMenuExpanded={isMenuExpanded}>
-              <IoStar />
-              <span>Favorites</span>
-            </MenuIcon>
-          </MenuItem>
-          <MenuItem disabled>
-            <MenuIcon isMenuExpanded={isMenuExpanded}>
-              <IoOptions />
-              <span>Options</span>
-            </MenuIcon>
-          </MenuItem>
-          <MenuItem>
-            {authenticated ? (
-              <MenuLink
-                to="/login"
-                exact
-                onClick={deAuthenticate}
-                data-testid="button-logout"
-              >
-                <MenuIcon isMenuExpanded={isMenuExpanded}>
-                  <IoExitOutline />
-                  <span>Logout</span>
-                </MenuIcon>
-              </MenuLink>
-            ) : (
-              <MenuLink to="/login" exact>
-                <MenuIcon isMenuExpanded={isMenuExpanded}>
-                  <IoEnterOutline />
-                  <span>Login</span>
-                </MenuIcon>
-              </MenuLink>
-            )}
-          </MenuItem>
-        </MenuList>
-      </Menu>
+      <Menu isExpanded={isMenuExpanded} />
       <MainContent data-testid="main-content">{children}</MainContent>
     </Container>
   );
